feat: add unpacker for bool
The reference implementation does not have such method on the unpacker.
But the generated netcode also uses GetInt() for boolean values.
This means that it could technically consume more than one byte if
it is an integer with extension bits set.

The original teeworlds protocol defines only 1 and 0 as valid values.
So technically anything else should be dropped.

But we still have to return a value from the method in the error case.
I decided to go with `false` you could also argue that in C
any value other than `0` should be considered `true`.