refactor: rename MessageKind to DDNetMessageCategory
Most of the other Kind enums are used to switch an enum.
In a next refactor I plan to use the name MessageKind for the kind of
a message ... duh.
The concept system or not sadly has no name in the teeworlds code.
Its either a bool true or false or a flag being set or not.
So I came up with the name Category. Because that seems to make the most
sense to me.
There are two categories: one for engine low level messages and one for
game related messages.

This frees up MessageKind for the different kinds of specific game and system
messages.

Also the enum values get the DDNET_ prefix to avoid naming collisions.