refactor: rename chunk kind to message kind
During the conversation in https://github.com/MilkeeyCat/ddnet_protocol/pull/108
It became clear some kind of struct that bundles the type of net message
and the union holding it would come in handy.

It now uses the clearer name DDNetMessageKind instead of ChunkKind.
Because chunk kind was the wrong name anyways. Chunk kind more refers
to a vital or non vital chunk. Not if it holds a sys.info message or a
sys.rcon_cmd message. The payload does not determine the type of chunk.

The refactor also introduced the proper DDNet and DDNET_ name prefixes
to avoid collisions.

And exposes the previously anonymous message union. Which allows for
cleaner code. Also on the user side.